separate info to another paper
