+++ src/scenes/Wallet/components/Balances/styles.js
@@ -42,9 +42,6 @@ const styles = (theme) => ({
     marginRight: theme.padding.unit.px,
     borderRadius: 2,
   },
-  importantNoteContainer: {
-    paddingBottom: theme.padding.xs.px,
-  },
 });
 
 export default styles;
